使用 State 接口和状态成员变量对 View 进行 UI 更改

+++ app/src/main/java/com/android/mvvm/repository/RepoRepositoryImpl.kt
@@ -1,42 +1,52 @@
 package com.android.mvvm.repository
 
-import com.android.mvvm.core.extension.tryCatchException
-import com.android.mvvm.core.model.DataState
-import com.android.mvvm.data.RepoBean
+import com.android.mvvm.core.base.ErrorState
+import com.android.mvvm.core.base.FailureState
+import com.android.mvvm.core.base.LoadingState
+import com.android.mvvm.core.base.State
+import com.android.mvvm.core.model.*
 import com.android.mvvm.data.dao.RepoBeanDao
+import com.android.mvvm.viewmodel.RepoViewModel.RepoState.CheckSuccess
+import com.android.mvvm.viewmodel.RepoViewModel.RepoState.RepoDataState
 import com.android.mvvm.web.api.RepoApi
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.flow
 import javax.inject.Inject
+
+interface RepoRepository {
+
+    suspend fun loadRepos(username: String): Flow<State>
+
+    suspend fun getRepos(): Flow<State>
+}
 
 class RepoRepositoryImpl @Inject constructor(
     private val repoApi: RepoApi,
     private val repoDao: RepoBeanDao
 ) : RepoRepository {
 
-    override suspend fun loadRepos(username: String): Flow<DataState<List<RepoBean>>> = flow {
-        try {
-            emit(DataState.Loading)
-            val results = repoApi.getRepos(username)
+    override suspend fun loadRepos(username: String): Flow<State> = flow {
+        emit(LoadingState)
+        val results = repoApi.getRepos(username)
+        if(results.isNotEmpty()){
             repoDao.deleteAllRepos()
             repoDao.insertAll(results)
-            emit(DataState.Success(results))
-        } catch (e: Exception) {
-            e.tryCatchException()
-            emit(DataState.Error(e))
+            emit(CheckSuccess(true))
+        }else{
+            emit(FailureState(-1,"该用户还没上传 Repo"))
         }
+    }.catch { e ->
+        emit(ErrorState(e))
     }
 
 
-    override suspend fun getRepos(): Flow<DataState<List<RepoBean>>> = flow {
-        try {
-            emit(DataState.Loading)
-            val results = repoDao.getAllRepos()
-            emit(DataState.Success(results))
-        } catch (e: Exception) {
-            e.tryCatchException()
-            emit(DataState.Error(e))
-        }
+    override suspend fun getRepos(): Flow<State> = flow {
+        emit(LoadingState)
+        val results = repoDao.getAllRepos()
+        emit(RepoDataState(results))
+    }.catch { e ->
+        emit(ErrorState(e))
     }
 
 }
